cofiguration firebase for android
