+++ pubspec.yaml
@@ -31,6 +31,7 @@ dependencies:
   google_maps_flutter: ^2.0.6
   location: ^4.2.3
   provider: ^5.0.0
+  cloud_firestore: ^2.2.2
 
 dev_dependencies:
